Version: 2013-02-14 ksh93v-
Sourced from:
5ac6216ba49a7950094d91aedca8c6df0ca0dbb7fcc8a62dc8eda6953a245f3f  ast-open.2012-11-21.tgz
2b9a3cd114182bc9a9ebc6efc684429d912eddbcd94ce50d8ef8eefa5018bc95  INIT.2013-02-14.tgz
1bfae3fb1d4b34c26fd6fc6f9191e5e54f7d41682e7d1d1aad30db4d738c58a2  ast-ksh.2013-02-14.tgz

+++ src/lib/libast/tm/tmxdate.c
@@ -135,7 +135,7 @@ powerize(Tm_t* tm, unsigned long p, unsigned long q, unsigned long u)
 		q *= 10;
 		t *= 10;
 	}
-	tm->tm_nsec += (int)(t % TMX_RESOLUTION);
+	tm->tm_nsec += (int)((unsigned long)t % TMX_RESOLUTION);
 	tm->tm_sec += (int)(t / TMX_RESOLUTION);
 }
